Contoh Program

+++ aljabar.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main(){
 
-	float x;
+	int x;
 	
 	cout << "f(x) = 2x^3 + 5x^2 - 7x + 10\n\n";
